pyproject.toml
[project]
name = "c0d3r"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-shell-plus>=1.1.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
